cleanup auth.tests
